roles y permisos gatess

+++ app/Http/Controllers/Admin/RoleController.php
@@ -7,6 +7,7 @@ use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use App\Permisos\Models\Role;
 use App\Permisos\Models\Permission;
+use Illuminate\Support\Facades\Gate;
 
 class RoleController extends Controller
 {
@@ -21,6 +22,9 @@ class RoleController extends Controller
      */
     public function index(Request $request)
     {
+        
+        Gate::authorize('haveaccess','admin.role.index');
+
         $nombre = $request->get('nombre'); 
         $roles = Role::where('nombre','like',"%$nombre%")->orderBy('nombre')->paginate(5);   
         return view('admin.role.index',compact('roles'));
@@ -32,7 +36,8 @@ class RoleController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function create()
-    {
+    {        
+        Gate::authorize('haveaccess','admin.role.create');
         $permisos = Permission::get(); 
         return view('admin.role.create',compact('permisos'));
     }
@@ -45,7 +50,7 @@ class RoleController extends Controller
      */
     public function store(Request $request)
     {
-        
+        Gate::authorize('haveaccess','admin.role.create');
         $request->validate([
             'nombre' => 'required|max:50|unique:roles,nombre',
             'slug' => 'required|max:50|unique:roles,slug',
@@ -73,6 +78,7 @@ class RoleController extends Controller
      */
     public function show(Role $role)
     { 
+        Gate::authorize('haveaccess','admin.role.show');
         $permission_role = [];
         foreach($role->permissions as $permission){
             $permission_role[] = $permission->id; 
@@ -90,6 +96,7 @@ class RoleController extends Controller
      */
     public function edit(Role $role)
     {
+        Gate::authorize('haveaccess','admin.role.edit');
         $permission_role = [];
         foreach($role->permissions as $permission){
             $permission_role[] = $permission->id; 
@@ -108,6 +115,7 @@ class RoleController extends Controller
      */
     public function update(Request $request, Role $role)
     { 
+        Gate::authorize('haveaccess','admin.role.edit');
         $request->validate([
             'nombre' => 'required|max:50|unique:roles,nombre,'.$role->id,
             'slug' => 'required|max:50|unique:roles,slug,'.$role->id,
@@ -115,8 +123,16 @@ class RoleController extends Controller
             'acceso-total' => 'required|in:si,no'
         ]);
 
-        $role = Role::update($request->all());
+        
+        
 
+ 
+        $role->fill($request->all())->save();
+
+        /*$role = Role::update(
+            
+            $request->all());
+*/
         if($request->get('permisos')){
             $role->permissions()->sync($request->get('permisos'));
         }
@@ -135,6 +151,7 @@ class RoleController extends Controller
      */
     public function destroy($id)
     {
+        Gate::authorize('haveaccess','admin.role.destroy');
         $rol = Role::findOrFail($id); 
         $rol->delete();
         return redirect()->route('admin.role.index')->with('datos','Registro eliminado correctamente');
